* updates to tspike format
git-svn-id: file:///home/jonas/soma/svn-restore.20080710/software/network/trunk@2842 5dccc5d9-beb7-0310-bf71-cc18d937a3f4

+++ src/tspike.h
@@ -16,7 +16,7 @@ const int TSPIKEWAVE_LEN = 32;
 struct TSpikeWave_t {
   uint8_t srcchan; 
   uint8_t valid; 
-  uint16_t filtid; 
+  uint32_t filtid; 
   int32_t threshold; 
   int32_t wave[TSPIKEWAVE_LEN]; 
 };
